pkging stands for packaging.

 Conflicts:
	CHANGELOG.md
	setup.data
	src/oclaunch.ml

Setup.data was removed since it create build probelms on other computers.
CHANGELOG just kept the two parts.
Oclaunch.ml file got some code to get execution time.

+++ release.sh
@@ -7,8 +7,16 @@
 
 # Record changes
 git commit -a -m "Version $(cat ./VERSION)"
-git tag -s v$(cat VERSION) -m "Release $(cat VERSION)"
+git tag -s v$(cat VERSION) -m "Release version $(cat VERSION)"
 
-# Cleanup dist directory to put the new archives
-rm -r dist/*
+# Cleanup dist directory to put the new archives, if exists
+if [ -d dist ]; then
+  rm -r dist/*
+fi
+# Binary archives
 ./0install.sh
+# Source code
+./pkg.sh
+
+# Signing everything
+./sign-dist.sh
